house-design-app/requirements.txt
numpy
matplotlib
Pillow
scipy
pytest